@@ -3,27 +3,31 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "scripts": {
+    "build": "yarn run build:prod",
+    "build:prod": "webpack --mode production",
+    "build:dev": "webpack --mode development"
+  },
   "dependencies": {
     "moment": "^2.11.2",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
+    "prop-types": "^15.7.2",
+    "react": "16.11.0",
+    "react-dom": "16.11.0",
     "react-redux": "^4.4.0",
-    "redux": "^3.3.1",
-    "redux-thunk": "^1.0.3",
-    "webpack": "^1.12.14"
+    "redux": "^3.7.2",
+    "redux-thunk": "^1.0.3"
   },
   "devDependencies": {
-    "babel-core": "^6.4.5",
+    "@babel/core": "^7.15.0",
+    "@babel/preset-env": "^7.15.0",
+    "@babel/preset-react": "^7.14.5",
     "babel-eslint": "^4.1.8",
-    "babel-loader": "^6.2.1",
+    "babel-loader": "^8.2.2",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
-    "eslint": "^1.10.3",
-    "eslint-loader": "^1.2.1",
-    "eslint-plugin-react": "^3.16.1",
     "path": "^0.12.7",
-    "watch": "^0.17.1",
-    "webpack": "^1.12.12",
-    "webpack-cli": "^3.1.1"
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.12",
+    "yarn": "1.22.11"
   }
 }
