@@ -9,10 +9,10 @@ import ConnectStatus from './ConnectStatus.jsx';
 const createStoreWithMiddleware = applyMiddleware(thunkMiddleware)(createStore);
 const store = createStoreWithMiddleware(reducer);
 
-const MountableApp = React.createClass({
-    render: function() {
+const MountableApp = class StatusPageApp extends React.Component {
+    render() {
         return <Provider store={store}><ConnectStatus dispatch={store.dispatch}/></Provider>
     }
-});
+}
 
 export default MountableApp;
